backend/requirements.txt
numpy>=1.24
shapely>=2.0
pillow>=10.0
rasterio>=1.3
requests>=2.31
python-dateutil>=2.8

python-dotenv>=1.0
